Serve the websockets on a self contained server
Given the websockets are going to use their own TLS subsystem it decomplicates things for them to be running on a seperate TCP listener.

Unfortunately it adds yet another moving piece that needs to be kept alive - possibly it would be better for the vending handler and the web handler to be compeletly different processes?