Add a container Machine that wraps over Hardware
This currently just makes the hardware threadsafe, but I plan to make it look after individual vending requests.

The idea is you ask the machine to start vending something and it returns you a futurey thing with an ID which you can query periodically from the client.
This will then quietly block in the background until the vend succeeds - complaining loudly if the server is shut down while it's active.